Completes the engine loop

+++ include/bbtypes.hpp
@@ -7,7 +7,7 @@
 
 namespace BlueBear {
 
-	using Tick = unsigned long;
+	using Tick = unsigned int;
 	using LuaReference = int;
 
 	enum TerrainType { TERRAIN_GRASS, TERRAIN_DIRT, TERRAIN_SNOW, TERAIN_SAND, TERRAIN_MOON, TERRAIN_UNDERGROUND };
